+++ dataproc_jupyter_plugin/controllers/vertex.py
@@ -19,24 +19,26 @@ import tornado
 from jupyter_server.base.handlers import APIHandler
 
 from dataproc_jupyter_plugin import credentials
-from dataproc_jupyter_plugin.services.vertex import vertex
+from dataproc_jupyter_plugin.services import vertex
 
 
-class SharedNetworkController(APIHandler):
+class UIConfigController(APIHandler):
     @tornado.web.authenticated
     async def get(self):
-        """Returns shared network"""
+        """Returns available ui config"""
         try:
+            region_id = self.get_argument("region_id")
             async with aiohttp.ClientSession() as client_session:
                 client = vertex.Client(
                     await credentials.get_cached(), self.log, client_session
                 )
-                network = await client.get_shared_network()
+
+                configs = await client.list_uiconfig(region_id)
                 response = []
-                for item in network:
-                    env = item.dict()
+                for config in configs:
+                    env = config.dict()
                     response.append(env)
                 self.finish(json.dumps(response))
         except Exception as e:
-            self.log.exception(f"Error fetching network shared from host: {str(e)}")
+            self.log.exception(f"Error fetching ui config: {str(e)}")
             self.finish({"error": str(e)})
