+++ dataproc_jupyter_plugin/services/compute.py
@@ -12,7 +12,6 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from google.cloud import storage
 from dataproc_jupyter_plugin import urls
 from dataproc_jupyter_plugin.commons.constants import (
     CONTENT_TYPE,
@@ -39,37 +38,6 @@ class Client:
             "Content-Type": CONTENT_TYPE,
             "Authorization": f"Bearer {self._access_token}",
         }
-
-    async def list_uiconfig(self, region_id):
-        try:
-            uiconfig = []
-            api_endpoint = f"https://{region_id}-aiplatform.googleapis.com/ui/projects/{self.project_id}/locations/{region_id}/uiConfig"
-
-            headers = self.create_headers()
-            async with self.client_session.get(
-                api_endpoint, headers=headers
-            ) as response:
-                if response.status == 200:
-                    resp = await response.json()
-                    if not resp:
-                        return uiconfig
-                    else:
-                        for machineconfig in resp.get("notebookRuntimeConfig").get("machineConfigs"):
-                            ramBytes_in_gb = round(int(machineconfig.get('ramBytes')) / 1000000000, 2)
-                            formatted_config = {
-                                "machineType": f"{machineconfig.get('machineType')} ({machineconfig.get('cpuCount')} CPUs, {ramBytes_in_gb} GB RAM)",
-                                "acceleratorConfigs": machineconfig.get("acceleratorConfigs")
-                            }
-                            uiconfig.append(formatted_config)
-                        return uiconfig
-                else:
-                    self.log.exception("Error listing ui config")
-                    raise Exception(
-                        f"Error getting vertex ui config: {response.reason} {await response.text()}"
-                    )
-        except Exception as e:
-            self.log.exception(f"Error fetching ui config: {str(e)}")
-            return {"Error fetching ui config": str(e)}
 
     async def list_region(self):
         try:
@@ -103,7 +71,9 @@ class Client:
         try:
             networks = []
             vertex_url = await urls.gcp_service_url(COMPUTE_SERVICE_NAME)
-            api_endpoint = f"{vertex_url}/compute/v1/projects/{self.project_id}/global/networks"
+            api_endpoint = (
+                f"{vertex_url}/compute/v1/projects/{self.project_id}/global/networks"
+            )
 
             headers = self.create_headers()
             async with self.client_session.get(
@@ -179,16 +149,3 @@ class Client:
         except Exception as e:
             self.log.exception(f"Error fetching shared network: {str(e)}")
             return {"Error fetching network shared from host": str(e)}
-
-    async def get_csb(self):
-        try:
-            cloud_storage_buckets = []
-            storage_client = storage.Client()
-            buckets = storage_client.list_buckets()
-            for bucket in buckets:
-                cloud_storage_buckets.append(bucket.name)
-            return cloud_storage_buckets
-
-        except Exception as e:
-            self.log.exception(f"Error fetching cloud storage buckets: {str(e)}")
-            return {"Error fetching cloud storage buckets": str(e)}
